refactor: move search related components
